첫번째 기본 기능 수정

+++ Homework/Num_SumAvr.cpp
@@ -12,7 +12,7 @@ int Add(int arr[], int len)
 	return result;
 }
 
-int Average(int arr[], int len)
+double Average(int arr[], int len)
 {
 	int sum = Add(arr, len);
 	return sum / len;
